# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: School Grade Analyzer API
servers:
  - url: /api
...
paths:
...
  /subscriptions/my:
    get:
      operationId: getMySubscription
      tags: [subscriptions]
      summary: Get current user's subscription
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
      responses:
        "200":
          description: Current subscription
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MySubscription"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /subscriptions/activate:
    post:
      operationId: activateSubscription
      tags: [subscriptions]
      summary: Activate a subscription plan (demo payment)
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActivateSubscriptionBody"
      responses:
        "200":
          description: Subscription activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MySubscription"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /grades/upload:
    post:
      operationId: uploadGrades
      tags: [grades]
      summary: Upload Excel file for grade analysis
      parameters:
        - $ref: "#/components/parameters/AuthorizationSessionHeader"
        - $ref: "#/components/parameters/SessionCookie"
      responses:
        "200":
          description: Successfully analyzed grades
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GradeAnalysisResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Subscription required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: "Opaque session token — Bearer <sid>."
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    MySubscription:
      type: object
      required: [plan, schoolMode, activatedAt]
      properties:
        plan:
          type: string
        schoolMode:
          type: string
        activatedAt:
          type: string
        expiresAt:
          type: ["string", "null"]
    ActivateSubscriptionBody:
      type: object
      required: [plan, schoolMode]
      properties:
        plan:
          type: string
          enum: [gratuit, standard, pro, max]
        schoolMode:
          type: string
          enum: [cem, lycee]
    StudentResult:
      type: object
      properties:
        name:
          type: string
        subjects:
          type: object
          additionalProperties:
            type: number
        average:
          type: number
        passed:
          type: boolean
        rank:
          type: integer
      required: [name, subjects, average, passed, rank]
    GradeAnalysisResult:
      type: object
      properties:
        students:
          type: array
          items:
            $ref: "#/components/schemas/StudentResult"
        summary:
          $ref: "#/components/schemas/GradeSummary"
        fileName:
          type: string
        totalStudents:
          type: integer
        schoolMode:
          type: string
        subjects:
          type: array
          items:
            type: string
      required: [students, summary, fileName, totalStudents, schoolMode, subjects]
    GradeSummary:
      type: object
      properties:
        classAverage:
          type: number
        highestAverage:
          type: number
        lowestAverage:
          type: number
        passRate:
          type: number
        passCount:
          type: integer
        failCount:
          type: integer
        topStudent:
          type: string
        weakestStudent:
          type: string
      required:
        [classAverage, highestAverage, lowestAverage, passRate, passCount, failCount, topStudent, weakestStudent]